+++ CHANGELOG.md
@@ -13,6 +13,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
   is asynchronous now, it has additional `onOpen` parameter
 - [WalletConnectCore#deleteSessionInternal](walletconnect/src/main/java/walletconnect/WalletConnectCore.kt) method
   is `private` now
+- [CallbackData#simplifiedName](walletconnect-core/src/main/java/walletconnect/core/session/callback/CallbackData.kt)
+  has `withContent:Boolean` parameter to get detailed message. 
+- Overridden `FailureType.toString` & `SignType.toString` methods
 
 ## [0.6.3] - 2022-22-08